Removed old map code and renamed MapReact -> ZoneMap
