--- web/src/global.js
@@ -1,5 +1,0 @@
-// TODO: Get rid of this once the map is moved to React.
-// See https://github.com/tmrowco/electricitymap-contrib/issues/2309
-const global = { zoneMap: null };
-
-export default global;
