+++ web/src/layout/main.js
@@ -30,7 +30,7 @@ import OnboardingModal from '../components/onboardingmodal';
 import LoadingOverlay from '../components/loadingoverlay';
 import Toggle from '../components/toggle';
 
-import MapReact from '../components/mapreact';
+import ZoneMap from '../components/zonemap';
 import MapLayer from '../components/maplayer';
 import ExchangeLayer from '../components/layers/exchangelayer';
 import SolarLayer from '../components/layers/solarlayer';
@@ -88,12 +88,11 @@ const Main = ({
           <LoadingOverlay visible={showLoadingOverlay} />
           <LeftPanel />
           <div id="map-container" className={location.pathname !== '/map' ? 'small-screen-hidden' : ''}>
-            <div id="zones" className="map-layer" style={{ display: 'none' }} />
-            <MapReact>
+            <ZoneMap>
               <MapLayer component={ExchangeLayer} />
               <MapLayer component={WindLayer} />
               <MapLayer component={SolarLayer} />
-            </MapReact>
+            </ZoneMap>
             <div id="watermark" className={`watermark small-screen-hidden ${brightModeEnabled ? 'brightmode' : ''}`}>
               <a href="http://www.tmrow.com/mission?utm_source=electricitymap.org&utm_medium=referral&utm_campaign=watermark" target="_blank">
                 <div id="built-by-tomorrow" />
